+++ test/DatasetBase/TranslatableIntegerTest.php
@@ -142,9 +142,11 @@ final class TranslatableIntegerTest extends TestCase
     #[DataProvider('provideInvalidTestCases')]
     public function testInsertOtherTypeInTypeCollection($item): void
     {
+        static::expectException(\TypeError::class);
+        static::expectExceptionMessageMatches('/must be of type int/');
+
         $translatable = new TranslatableInteger();
         $translatable->setTranslation('en-GB', $item);
-        static::assertCount(0, $translatable->getLocaleKeys());
     }
 
     public function testSerialization(): void
@@ -168,9 +170,10 @@ final class TranslatableIntegerTest extends TestCase
     #[DataProvider('provideInvalidTestCases')]
     public function testInvalidFallback($anyValue): void
     {
-        $translatable = new TranslatableInteger();
+        static::expectException(\TypeError::class);
+        static::expectExceptionMessageMatches('/must be of type int/');
 
+        $translatable = new TranslatableInteger();
         $translatable->setFallback($anyValue);
-        static::assertNull($translatable->getFallback());
     }
 }